some code refactor / clean up , add Computer.makeMove function
